scripts/vscode-extensions/requirements.txt
PyYAML==6.0
beautifulsoup4==4.11.1
python-markdown-maker==1.0
requests==2.28.1
typing_extensions==4.2.0